# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RESTful API - Yu-Gi-Oh!",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://agree.test:8080"
        }
    ],
    "paths": {
...
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Lista Categorias y SubCategorias",
                "description": "",
                "operationId": "887f1ab441defeefa3a9afb3d8bae769",
                "parameters": [
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "application/json",
                        "required": true
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token obtenido al iniciar sesion",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Crear Categorias y SubCategorias",
                "description": "El parametro parent_code es opcional, si se desea crear una categoria padre no se envia en el body.",
                "operationId": "528e99ff0e3d24de9a9870ab99619039",
                "parameters": [
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "application/json",
                        "required": true
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token obtenido al iniciar sesion",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "sku_id": {
                                        "type": "string"
                                    },
                                    "parent_code": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Mostrar Categoria y SubCategoria",
                "description": "",
                "operationId": "a604c1a3cf1dcb0fc186ff5d0a30779b",
                "parameters": [
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "application/json",
                        "required": true
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token obtenido al iniciar sesion",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Actualizar Categorias y SubCategorias",
                "description": "",
                "operationId": "fa1a3d6ff8756f8f955dac873b8c9d13",
                "parameters": [
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "application/json",
                        "required": true
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token obtenido al iniciar sesion",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "sku_id": {
                                        "type": "string"
                                    },
                                    "parent_id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Eliminar Categorias y SubCategorias",
                "description": "",
                "operationId": "f2c692a4f491c2cc38650f93ee76aa60",
                "parameters": [
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "application/json",
                        "required": true
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token obtenido al iniciar sesion",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        }
    },
...
    "security": [
        []
    ]
}